protocols login doc, proofs moved to own file, checks of the protocol
