+++ src/protocols/login.js
@@ -4,6 +4,7 @@ const utils = require('../utils');
 const claim = require('../claim/claim');
 const CONSTANTS = require('../constants');
 const Entry = require('../claim/entry/entry');
+const proofs = require('./proofs');
 
 const mtHelpers = require('../sparse-merkle-tree/sparse-merkle-tree-utils');
 const smt = require('../sparse-merkle-tree/sparse-merkle-tree');
@@ -56,7 +57,7 @@ const signPacket = function signPacket(signatureRequest, usrAddr, kc, ksign, pro
 };
 */
 
-const signIdenAssertV01 = function signIdenAssertV01(signatureRequest, ethAddr, ethName, kc, ksign, proofOfKSign, proofOfEthName, expirationTime) {
+const signIdenAssertV01 = function signIdenAssertV01(signatureRequest, ethAddr, ethName, proofOfEthName, kc, ksign, proofOfKSign, expirationTime) {
   const date = new Date();
   const currentTime = Math.round((date).getTime() / 1000);
   const jwsHeader = {
@@ -98,10 +99,17 @@ const signIdenAssertV01 = function signIdenAssertV01(signatureRequest, ethAddr, 
 const verifyIdenAssertV01 = function verifyIdenAssertV01(jwsHeader, jwsPayload, signatureBuffer) {
   // TODO check data structure scheme
 
+  // check if jwsHeader.alg is iden3.sig.v0_1 (SIGALGV01)
   if (jwsHeader.alg !== SIGALGV01) {
     return false;
   }
 
+  // check that jwsPayload.proofOfKSign.proofs.length < 2
+  if (jwsPayload.proofOfKSign.proofs.length>2) {
+    return false;
+  }
+
+  // check times iat < current < exp
   const date = new Date();
   const current = Math.round((date).getTime() / 1000);
   if (!((jwsHeader.iat <= current) && (current <= jwsHeader.exp))) {
@@ -109,8 +117,9 @@ const verifyIdenAssertV01 = function verifyIdenAssertV01(jwsHeader, jwsPayload, 
     return false;
   }
 
-  const ksign = ''; // TODO get ksign from jwsPayload.proofOfKSign.Leaf
+  // TODO check jwsPayload.ksign with jwsPayload.proofOfKSign.Leaf
 
+  // check verify signature with jwsPayload.ksign
   // as verifying a signature is cheaper than verifying a merkle tree proof, first we verify signature with ksign
   const header64 = Buffer.from(JSON.stringify(jwsHeader)).toString('base64');
   const payload64 = Buffer.from(JSON.stringify(jwsPayload)).toString('base64');
@@ -125,22 +134,19 @@ const verifyIdenAssertV01 = function verifyIdenAssertV01(jwsHeader, jwsPayload, 
 
   // TODO verify that signature is by jwsHeader.iss
 
-  // TODO verify proofOfEthName
-
-  // verify proofOfKSign
-  if (jwsPayload.proofOfKSign.proofs.length != 2) {
-    console.trace();
-    return false;
-  }
-  // if (jwsPayload.proofOfKSign.ethaddrs.length !== jwsPayload.proofOfKSign.proofs.length) {
-  //   return false;
-  // }
-
-  // Use verifyProofClaimFull to verify proofOfKSign
   const relayAddr = '0xe0fbce58cfaa72812103f003adce3f284fe5fc7c';
-  jwsPayload.proofOfKSign.proofs = jwsPayload.proofOfKSign.proofs;
-  if (!verifyProofClaimFull(jwsPayload.proofOfKSign, relayAddr)) {
-    console.trace();
+
+  // TODO check that jwsPayload.form.proofOfEthName == jwsPayload.form.ethName
+
+  // verify proofOfEthName
+  if (!proofs.verifyProofClaimFull(jwsPayload.form.proofOfEthName.proofOfClaimAssignName, relayAddr)) {
+ 	return false; 
+  }
+
+  // TODO check that jwsPayload.proofOfKSign ksign == ksign
+
+  // check jwsPayload.proofOfKSign
+  if (!proofs.verifyProofClaimFull(jwsPayload.proofOfKSign, relayAddr)) {
     return false;
   }
 
@@ -171,161 +177,6 @@ const verifySignedPacket = function verifySignedPacket(signedPacket) {
   return verified;
 };
 
-// for general purposes
-
-/** Class representing proof of a valid claim, from the claim to the top level
- * tree root, including proofs of non-existence of the revoked/next version of
- * each claim.  Includes the top level tree root key signature by the owner of
- * the top level tree. */
-class ProofClaimFull {
-  /**
-   * Create a ClaimProof
-   * @param {Signature} rootKeySig
-   * @param {Number} date
-   * @param {Leaf} leaf
-   * @param {[]ProofClaim} proofs
-   */
-  constructor(rootKeySig, date, leaf, proofs) {
-    this.signature = rootKeySig;
-    this.date = date;
-    this.leaf = leaf;
-    this.proofs = proofs;
-  }
-}
-
-/**
- * Class representing a merkle tree proof of existence of a leaf, a merkle tree
- * proof of non existence of the same leaf with the following version */
-class MtpProof {
-  /**
-   * Create an MtpProof
-   * @param {MerkleTreeProof} mtp0
-   * @param {MerkleTreeProof} mtp1
-   * @param {key} root
-   * @param {SetRootAux|null} aux
-   */
-  constructor(mtp0, mtp1, root, aux) {
-    this.mtp0 = mtp0;
-    this.mtp1 = mtp1;
-    this.root = root;
-    this.aux = aux;
-  }
-}
-
-/**
- * Auxiliary data required to build a set root claim
- */
-class SetRootAux {
-  constructor(ethAddr, version, era) {
-    this.ethAddr = ethAddr;
-    this.ver = version;
-    this.era = era;
-  }
-}
-
-// TODO: Move this to claim utils
-const incClaimVersion = function incClaimVersion(claim) {
-  //let entry = new Entry();
-  //entry.fromHexadecimal(claim);
-  const version = claim.elements[3].slice(20, 24).readUInt32BE(0);
-  claim.elements[3].writeUInt32BE(version + 1, claim.elements[3].length - 64 / 8 - 32 / 8);
-}
-
-// TODO: Move this to merkle-tree utils
-const isMerkleTreeProofExistence = function isMerkleTreeProofExistence(proofHex) {
-  const proofBuff = mtHelpers.parseProof(proofHex);
-  const flagNonExistence = mtHelpers.getBit(proofBuff.flagExistence, 0);
-  return !flagNonExistence;
-}
-
-/**
- * Verify a ProofClaimFull from the claim to the blockchain root
- * @param{ProofClaimFull} proof
- */
-const verifyProofClaimFull = function verifyProofClaimFull(proof, relayAddr) {
-  // Verify that signature(proof.proofs[proof.proofs.length - 1].root) === proof.rootKeySig
-  let rootK = proof.proofs[proof.proofs.length - 1].root;
-  if (rootK.substr(0, 2) !== '0x') {
-    rootK = '0x' + rootK;
-  }
-  // const date = proof.proofs[proof.proofs.length - 1].Date;
-  const date = proof.date;
-  const dateBytes = utils.uint64ToEthBytes(date);
-  const dateHex = utils.bytesToHex(dateBytes);
-  const msg = `${rootK}${dateHex.slice(2)}`;
-  const msgBuffer = ethUtil.toBuffer(msg);
-  const msgHash = ethUtil.hashPersonalMessage(msgBuffer);
-  const msgHashHex = utils.bytesToHex(msgHash);
-  let sig = utils.hexToBytes(proof.signature);
-  sig[64] += 27;
-  const signatureHex = utils.bytesToHex(sig);
-  if (!utils.verifySignature(msgHashHex, signatureHex, relayAddr)) { // mHex, sigHex, addressHex
-    // if (!utils.verifySignature(msg, proof.rootSig, relayAddr)) {  mHex, sigHex, addressHex
-    console.trace();
-    return false;
-  }
-
-  // For now we only allow proof verification of Nameserver (one level) and
-  // Relay (two levels: relay + user)
-  if (proof.proofs.length > 2 || proof.proofs.length < 1) {
-    console.trace();
-    return false;
-  }
-
-  let leaf = new Entry();
-  leaf.fromHexadecimal(proof.leaf);
-  var rootKey = '';
-  for (var i = 0; i < proof.proofs.length; i++) {
-    // for (var i = proof.proofs.length-1; i>=0; i--) {
-    mtpEx = proof.proofs[i].mtp0
-    mtpNoEx = proof.proofs[i].mtp1
-    // WARNING: leafNoEx points to the same content of leaf, so modifying leafNoEx modifies leaf!
-    //var leafNoEx = leaf
-    //incClaimVersion(leafNoEx)
-    rootKey = proof.proofs[i].root
-
-    if (!isMerkleTreeProofExistence(mtpEx)) {
-      console.trace(mtpEx);
-      return false;
-    }
-    // console.trace(leaf);
-    // console.trace(leaf.hi(), leaf.hv());
-    if (smt.checkProof(rootKey, mtpEx, utils.bytesToHex(leaf.hi()), utils.bytesToHex(leaf.hv())) !== true) {
-      console.log('rootKey: ' + rootKey);
-      console.log('proof: ' + mtpEx);
-      console.log('hi: ' + utils.bytesToHex(leaf.hi()));
-      console.log('hv: ' + utils.bytesToHex(leaf.hv()));
-      console.log('leaf: ' + leaf.toHexadecimal());
-      console.trace(leaf);
-      return false;
-    }
-    if (isMerkleTreeProofExistence(mtpNoEx)) {
-      console.trace();
-      return false;
-    }
-    incClaimVersion(leaf)
-    if (smt.checkProof(rootKey, mtpNoEx, utils.bytesToHex(leaf.hi()), utils.bytesToHex(leaf.hv())) !== true) {
-      console.trace();
-      return false;
-    }
-
-    if (i === proof.proofs.length - 1) {
-      break;
-    }
-    const version = proof.proofs[i].aux.ver;
-    const era = proof.proofs[i].aux.era;
-    const ethAddr = proof.proofs[i].aux.ethAddr;
-    // leaf = new iden3.claims.SetRootKey(version, era, ethAddr, rootKey);
-    leaf = new claim.Factory(CONSTANTS.CLAIMS.SET_ROOT_KEY.ID, {
-      version: version,
-      era: era,
-      id: ethAddr,
-      rootKey: rootKey
-    }).createEntry();
-  }
-
-  return true;
-}
 
 module.exports = {
   newRequestIdenAssert,
@@ -333,10 +184,6 @@ module.exports = {
   // signPacket,
   verifyIdenAssertV01,
   verifySignedPacket,
-  verifyProofClaimFull,
-  ProofClaimFull,
-  MtpProof,
-  SetRootAux
 };
 
 /*
@@ -352,7 +199,7 @@ verificar paquet firmat:
 		- verificar objecte del tipus iden3.sig.v0_1
 			- verificar JWS_HEADER.alg=='el que toca'
 		- verificar que JWS_PAYLOAD.proof_ksgin.proofs.length<=2
-		- verificar que JWS_PAYLOAD.proof_ksgin.ethaddrs.length=JWS_PAYLOAD.proof_ksgin.proofs.length - 1
+		// ja no, per la nova estructura de dades - verificar que JWS_PAYLOAD.proof_ksgin.ethaddrs.length=JWS_PAYLOAD.proof_ksgin.proofs.length - 1
 		- verificar paràmetres del JWS_HEADER
 			- iat < current < exp
 		- agafar ksign de JWS_PAYLOAD.proof_ksign.leaf
