+++ src/protocols/login.test.js
@@ -18,6 +18,7 @@ const ksign = kc.importKey(testPrivKHex);
 const usrAddr = '0x52dc5fa952194ad6c3268666fc4e64407a1d457a';
 const ethName = 'usertest@iden3.io';
 
+// following proofOfEthName and proofOfKSign are generated by go-iden3
 const proofOfEthName = {
   claimAssignName: '0x00000000000000000000000000000000000000000000000000000000000000000000000000000000000000007b471a1bdbd3b8ac98f3715507449f3a8e1f3b22008c8efcda9e563cf153563941b60fc5ac88336fc58d361eb0888686fadb99760000000000000000000000000000000000000000000000000000000000000003',
   ethAddr: '0x7b471a1bdbd3b8ac98f3715507449f3a8e1f3b22',
@@ -60,7 +61,7 @@ const proofOfKSign = {
   signature: '000c3f2ecd731905c8ce1e05de6a1edd09fe06611fef1cd700d3a84537bf6dc21e7b8f158f252cc583542c449b824cbf21080b9c5b46d27c036ceb32f51c2d2801'
 };
 
-describe('[protocol] login', () => {
+describe('[protocol] interns', () => {
   it('test bigint', () => {
     bigInt(8).toString();
     bigInt(8).shiftLeft(2);
@@ -88,14 +89,16 @@ describe('[protocol] login', () => {
   });
 
   it('verify ProofClaimFull (proofOfClaimAssignName & proofOfKSign)', () => {
-    const assignNameVerified = iden3.protocols.login.verifyProofClaimFull(proofOfEthName.proofOfClaimAssignName, relayAddr);
+    const assignNameVerified = iden3.protocols.verifyProofClaimFull(proofOfEthName.proofOfClaimAssignName, relayAddr);
     expect(assignNameVerified).to.be.equal(true);
 
-    const ksignVerified = iden3.protocols.login.verifyProofClaimFull(proofOfKSign, relayAddr);
+    const ksignVerified = iden3.protocols.verifyProofClaimFull(proofOfKSign, relayAddr);
     expect(ksignVerified).to.be.equal(true);
   });
+});
 
-  it('newRequestIdenAssert', () => {
+describe('[protocol] login flow', () => {
+  it('newRequestIdenAssert & signIdenAssertV01 & verifySignedPacket', () => {
     const date = new Date();
     const unixtime = Math.round((date).getTime() / 1000);
     const minutes = 20; // will be setted in global consts or in a config file
@@ -103,11 +106,9 @@ describe('[protocol] login', () => {
 
     const signatureRequest = iden3.protocols.login.newRequestIdenAssert('0xorigin', 'session01', timeout);
 
-    const res = iden3.protocols.login.verifyProofClaimFull(proofOfKSign, relayAddr);
-    expect(res).to.be.equal(true);
 
     const expirationTime = unixtime + (3600 * 60);
-    const signedPacket = iden3.protocols.login.signIdenAssertV01(signatureRequest, usrAddr, ethName, kc, ksign, proofOfKSign, proofOfEthName, expirationTime);
+    const signedPacket = iden3.protocols.login.signIdenAssertV01(signatureRequest, usrAddr, ethName, proofOfEthName, kc, ksign, proofOfKSign, expirationTime);
 
     const verified = iden3.protocols.login.verifySignedPacket(signedPacket);
     expect(verified).to.be.equal(true);
