+++ src/protocols/protocols.js
@@ -1,7 +1,9 @@
 
 const login = require('./login');
+const proofs = require('./proofs');
 
 module.exports = {
 	login,
+	verifyProofClaimFull: proofs.verifyProofClaimFull,
 };
